requirements.txt
altair==5.4.1
attrs==24.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
filelock==3.17.0
fsspec==2025.2.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
huggingface-hub==0.29.2
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.13.2
numpy==2.1.3
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
pyarrow==18.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
safetensors==0.5.3
selenium==4.27.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
streamlit==1.40.0
tenacity==9.0.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.1
transformers==4.49.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
websocket-client==1.8.0
wsproto==1.2.0
